Update CCTV_Self_made.py
I just started to explore git and I found your code and I tried executing it. It's really a good work and interesting too. And the only issue I faced was after pressing 'q' also gui was open and it was showing as not responding. So, I recommend you to call destoryWindow. If I'm wrong please ignore this.